Split core semantics tests
